BITの修正

+++ BIT.cpp
@@ -5,20 +5,20 @@ using namespace std;
 //1-indexdであることに注意する
 
 struct BIT {
-	vector<int> bit;
+	vector<LL> bit;
 	BIT(int n) {
 		bit.resize(n);
 		for (int i = 0; i < n; i++)bit[i] = 0;
 	}
 	int sum(int i) {
-		int s = 0;
+		LL s = 0;
 		while (i>0) {
 			s += bit[i];
 			i -= i& -i;
 		}
 		return s;
 	}
-	void add(int i, int x) {
+	void add(int i, LL x) {
 		while (i <= 202020) {
 			bit[i] += x;
 			i += i& -i;
